Update AjaxFileUpload.js
In server side, programmer usually use html input tag name to take data. In this script, I see file uploaded with name of file so that we have to use $_FILES.
in my code change it will be $_FILES['tag_input_name'].

When using this plugin, i have a problem when declaring ajaxfileupload twice with different input id and url target in one page. The url used for upload is the last declared. 

here are the example:
$(document).ready(function(){
$('#input-a').ajaxFileUpload({
			url: '/target_a/',
			autoUpload: true,
			dataType: 'json',
			onSuccess: function(data, status)
			{});

$('#input-b').ajaxFileUpload({
			url: '/target_b/',
			autoUpload: true,
			dataType: 'json',
			onSuccess: function(data, status)
			{});
);

when using file input-b, it runs well. But, if we upload with input-a it doesn't run because it uploads using the last declared url.

thanks for your plugin.